chore: moved widgets into respective directories
