+++ lib/community/widgets/post_card_view_comfortable.dart
@@ -8,7 +8,7 @@ import 'package:markdown/markdown.dart' hide Text;
 import 'package:thunder/community/enums/community_action.dart';
 import 'package:thunder/core/models/models.dart';
 import 'package:thunder/post/enums/post_action.dart';
-import 'package:thunder/post/widgets/post_action_bottom_sheet.dart';
+import 'package:thunder/post/widgets/post_bottom_sheet/post_action_bottom_sheet.dart';
 import 'package:thunder/community/widgets/post_card_actions.dart';
 import 'package:thunder/community/widgets/post_card_metadata.dart';
 import 'package:thunder/core/enums/media_type.dart';
