+++ lib/post/widgets/post_bottom_sheet/community_post_action_bottom_sheet.dart
@@ -9,7 +9,7 @@ import 'package:thunder/community/enums/community_action.dart';
 import 'package:thunder/core/models/models.dart';
 import 'package:thunder/feed/feed.dart';
 import 'package:thunder/post/enums/post_action.dart';
-import 'package:thunder/post/widgets/post_action_bottom_sheet.dart';
+import 'package:thunder/post/widgets/post_bottom_sheet/post_action_bottom_sheet.dart';
 import 'package:thunder/shared/bottom_sheet_action.dart';
 import 'package:thunder/shared/divider.dart';
 import 'package:thunder/thunder/thunder_icons.dart';
