+++ lib/post/widgets/post_bottom_sheet/post_action_bottom_sheet.dart
@@ -10,10 +10,10 @@ import 'package:thunder/community/widgets/post_card_metadata.dart';
 import 'package:thunder/core/enums/full_name.dart';
 import 'package:thunder/core/models/models.dart';
 import 'package:thunder/post/enums/post_action.dart';
-import 'package:thunder/post/widgets/community_post_action_bottom_sheet.dart';
-import 'package:thunder/post/widgets/general_post_action_bottom_sheet.dart';
+import 'package:thunder/post/widgets/post_bottom_sheet/community_post_action_bottom_sheet.dart';
+import 'package:thunder/post/widgets/post_bottom_sheet/general_post_action_bottom_sheet.dart';
 import 'package:thunder/instance/widgets/instance_action_bottom_sheet.dart';
-import 'package:thunder/post/widgets/post_post_action_bottom_sheet.dart';
+import 'package:thunder/post/widgets/post_bottom_sheet/post_post_action_bottom_sheet.dart';
 import 'package:thunder/shared/share/share_action_bottom_sheet.dart';
 import 'package:thunder/user/widgets/user_action_bottom_sheet.dart';
 import 'package:thunder/user/enums/user_action.dart';
