+++ lib/post/widgets/widgets.dart
@@ -1,10 +1,11 @@
-export 'community_post_action_bottom_sheet.dart';
-export 'general_post_action_bottom_sheet.dart';
-export 'post_action_bottom_sheet.dart';
+export 'post_bottom_sheet/community_post_action_bottom_sheet.dart';
+export 'post_bottom_sheet/general_post_action_bottom_sheet.dart';
+export 'post_bottom_sheet/post_action_bottom_sheet.dart';
 export 'post_card_title.dart';
-export 'post_metadata.dart';
+export 'post_body/post_metadata.dart';
 export 'post_page_app_bar.dart';
-export 'post_post_action_bottom_sheet.dart';
-export 'post_quick_actions_bar.dart';
+export 'post_bottom_sheet/post_post_action_bottom_sheet.dart';
+export 'post_body/post_quick_actions_bar.dart';
 export 'post_status_icon.dart';
-export 'post_view.dart';
+export 'post_body/post_body.dart';
+export 'post_body/post_body_preview.dart';
